+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`9230` rotki will now provide quick options for the date time picker.
 * :bug:`10960` rotki will now support swaps with fees in multiple different assets.
 * :bug: `10964` Fix an issue where the zoom state in the net value graph resets when the price is being queried.